requirements.txt
google-cloud-storage==1.19.1
google-cloud-pubsub==1.0.2
